New Components
Added Temp Component to show current Temperature.

Added Config component which has to be filled out before the MainMenue and the Well Tab is shown. (Only tested with a Button now)

Deleted useless Resources

To Do:
Don´t know how to set the Number of Circuits yet: Should I only make a TextBox to put the Number into and Jakob just gives me the WellIndex where which Sample is?
How to visualise where which Sample is:
WellController1 = red
WellControler2= green????
Change the Color of the TimeStamp?