Layout Changes
Changes in the Size of different Components so that the ControllMenu doesn´t look that chaotic. Made the Background color a little brighter.
Changed the Color of the BarGraph so that the grey is the same grey as on an disabled Button.
Deletet a useless native Statement that caused a warning.